+++ batch_test.go
@@ -1,3 +1,6 @@
+//go:build !app_engine
+// +build !app_engine
+
 // Copyright 2014-2022 Aerospike, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -29,6 +32,38 @@ import (
 
 // ALL tests are isolated by SetName and Key, which are 50 random characters
 var _ = gg.Describe("Aerospike", func() {
+
+	gg.Describe("BatchGetOperate operations", func() {
+		var ns = *namespace
+		var set = randString(50)
+
+		gg.It("must return the result with same ordering", func() {
+			for _, keyCount := range []int{256, 1} {
+				var keys []*as.Key
+				for i := 0; i < keyCount; i++ {
+					key, _ := as.NewKey(ns, set, i)
+					client.PutBins(nil, key, as.NewBin("i", i), as.NewBin("j", i))
+
+					keys = append(keys, key)
+				}
+
+				ops := []*as.Operation{as.GetBinOp("i"), as.PutOp(as.NewBin("h", 1))}
+				_, err := client.BatchGetOperate(nil, keys, ops...)
+				gm.Expect(err).To(gm.HaveOccurred())
+
+				ops = []*as.Operation{as.GetBinOp("i")}
+				recs, err := client.BatchGetOperate(nil, keys, ops...)
+
+				gm.Expect(err).ToNot(gm.HaveOccurred())
+				for i, rec := range recs {
+					gm.Expect(len(rec.Bins)).To(gm.Equal(1))
+					gm.Expect(rec.Bins["i"]).To(gm.Equal(i))
+				}
+
+			}
+		}) // it
+
+	}) // describe
 
 	gg.Describe("Batch Write operations", func() {
 		var ns = *namespace
@@ -255,6 +290,43 @@ var _ = gg.Describe("Aerospike", func() {
 				gm.Expect(op1.BatchRec().Record.Bins).To(gm.Equal(as.BinMap{"i": 0, "j": 5}))
 				gm.Expect(op1.BatchRec().InDoubt).To(gm.BeFalse())
 
+			})
+
+			gg.It("must successfully execute a BatchOperate for many keys", func() {
+				if *dbaas {
+					gg.Skip("Not supported in DBAAS environment")
+				}
+
+				gm.Expect(err).ToNot(gm.HaveOccurred())
+				bwPolicy := as.NewBatchWritePolicy()
+				bdPolicy := as.NewBatchDeletePolicy()
+
+				var keys []*as.Key
+				for i := 0; i < 64; i++ {
+					key, _ := as.NewKey(ns, set, i)
+					if i == 0 {
+						keys = append(keys, key)
+					}
+					bin0 := as.NewBin("count", i)
+					err := client.PutBins(nil, key, bin0)
+					gm.Expect(err).ToNot(gm.HaveOccurred())
+				}
+
+				for _, sendKey := range []bool{true, false} {
+					bwPolicy.SendKey = sendKey
+					bdPolicy.SendKey = sendKey
+					bpolicy.SendKey = !sendKey
+
+					var brecs []as.BatchRecordIfc
+					for _, key := range keys {
+						brecs = append(brecs, as.NewBatchWrite(bwPolicy, key, as.PutOp(as.NewBin("bin1", "a")), as.PutOp(as.NewBin("bin2", "b"))))
+						brecs = append(brecs, as.NewBatchDelete(bdPolicy, key))
+						brecs = append(brecs, as.NewBatchRead(nil, key, []string{"bin2"}))
+					}
+
+					err := client.BatchOperate(bpolicy, brecs)
+					gm.Expect(err).ToNot(gm.HaveOccurred())
+				}
 			})
 
 			gg.It("must successfully execute a delete op", func() {