+++ execute_task.go
@@ -180,7 +180,8 @@ func (etsk *ExecuteTask) grpcIsDone() (bool, Error) {
 
 	client := kvs.NewQueryClient(conn)
 
-	ctx, _ := context.WithTimeout(context.Background(), NewInfoPolicy().Timeout)
+	ctx, cancel := context.WithTimeout(context.Background(), NewInfoPolicy().Timeout)
+	defer cancel()
 
 	streamRes, gerr := client.BackgroundTaskStatus(ctx, &req)
 	if gerr != nil {
@@ -210,13 +211,5 @@ func (etsk *ExecuteTask) grpcIsDone() (bool, Error) {
 			etsk.clnt.returnGrpcConnToPool(conn)
 			return false, nil
 		}
-
-		if !res.GetHasNext() {
-			etsk.clnt.returnGrpcConnToPool(conn)
-			return false, nil
-		}
 	}
-
-	etsk.clnt.returnGrpcConnToPool(conn)
-	return true, nil
 }