+++ exists_command.go
@@ -133,7 +133,8 @@ func (cmd *existsCommand) ExecuteGRPC(clnt *ProxyClient) Error {
 
 	client := kvs.NewKVSClient(conn)
 
-	ctx := cmd.policy.grpcDeadlineContext()
+	ctx, cancel := cmd.policy.grpcDeadlineContext()
+	defer cancel()
 
 	res, gerr := client.Exists(ctx, &req)
 	if gerr != nil {