+++ read_command.go
@@ -294,7 +294,8 @@ func (cmd *readCommand) ExecuteGRPC(clnt *ProxyClient) Error {
 
 	client := kvs.NewKVSClient(conn)
 
-	ctx := cmd.policy.grpcDeadlineContext()
+	ctx, cancel := cmd.policy.grpcDeadlineContext()
+	defer cancel()
 
 	res, gerr := client.Read(ctx, &req)
 	if gerr != nil {