+++ server_command.go
@@ -125,7 +125,8 @@ func (cmd *serverCommand) ExecuteGRPC(clnt *ProxyClient) Error {
 
 	client := kvs.NewQueryClient(conn)
 
-	ctx := cmd.policy.grpcDeadlineContext()
+	ctx, cancel := cmd.policy.grpcDeadlineContext()
+	defer cancel()
 
 	streamRes, gerr := client.BackgroundExecute(ctx, &req)
 	if gerr != nil {