+++ touch_command.go
@@ -166,7 +166,8 @@ func (cmd *touchCommand) ExecuteGRPC(clnt *ProxyClient) Error {
 
 	client := kvs.NewKVSClient(conn)
 
-	ctx := cmd.policy.grpcDeadlineContext()
+	ctx, cancel := cmd.policy.grpcDeadlineContext()
+	defer cancel()
 
 	res, gerr := client.Touch(ctx, &req)
 	if gerr != nil {